refactor verifyFileExist and verifyFileExtension to FileIntegrityChecker
